Wait for network-online in systemd unit file
There's a long page in the systemd docs about when to use each of the 3
different network targets(!) and it looks like this is the right one:
https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

I don't think this fix will enable mDNS over USB (unless you manually
restart qmdns-responder after plugging it in). qmdns-responder binds to
0.0.0.0/::: (all interfaces), but I think once the socket is bound it
doesn't automatically start listening on new interfaces as well; it's
less "listen on all interfaces" and more "listen on all interfaces that
are up right now." The better fix is to have qmdns-responder watch for
new interfaces somehow and re-bind to 0.0.0.0 when a new one starts up.